requirements.txt
django-admin=3.0
social-auth-app-django
PyMySQL
decorator
numpy
requests